refactor: move default config to lib
